Adding failing and stale hosts in check output
Fixing array output and style guide cleanups

Change array check to use empty?

We only want to determine if the array is empty and has values,
whereas any? seems to check if anything evaluates to true, so if the
array is empty we get nilclass errors

Fixing the tests

Fixing formatting

Fixing tests with array ! string